Remove use of deprecated 'DisplayName' proprety
Apparently the 'DisplayName' property of the rectangle class was removed between the initial development of this project and the release of MATLAB R2015a (which is my version). Setting the 'DisplayName' property resulted in the error: "Error using rectangle There is no DisplayName property on the Rectangle class." Removing line 201 resolved this error.